Created "Game Over" Screen
It still needs some finishing touches. You can't run it as is in the project code. To test this class, I actually added it to the lab from 10/24/17 where we created a picture.
Ignore my comments (ie. section 1, 2, 3, ...); all of those were for math purposes to be able to create the letters. I calculated it using the pixel art I made in the last update and then used some proportions to adjust it to our project's canvas size.